Move TopLine into its own file
